requirements.txt
asgiref==3.8.1
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.1
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
PyJWT==2.10.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
